- review feedback comment - made plivoclient api compatible with existing code
